Fix login logic

+++ frontend/src/user/pages/Login/Login.jsx
@@ -3,6 +3,7 @@ import axios from "axios";
 import ServerRoute from "../../../ServerRoute";
 import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
+import Swal from "sweetalert2";
 
 function Login() {
   const navigate = useNavigate();
@@ -16,15 +17,34 @@ function Login() {
     e.preventDefault();
     const username = document.getElementById("username").value;
     const password = document.getElementById("password").value;
+
+    if (username === "" || password === "") {
+      Swal.fire({
+        icon: "error",
+        title: "Error",
+        text: "Field must not be empty"
+      });
+      return;
+    }
+
     axios.post(ServerRoute + "/login", {
       username,
       password
     }).then(function(r){
-      const jwtToken = r.data.other;
-      // console.log(jwtToken);
-      localStorage.setItem("posjwt", jwtToken);
-      // alert();
-      navigate("/", { replace: true });
+      // console.log(r);
+      if (r.data.status === "success") {
+        const jwtToken = r.data.other;
+        // console.log(jwtToken);
+        localStorage.setItem("posjwt", jwtToken);
+        // alert();
+        navigate("/", { replace: true });
+      } else {
+        Swal.fire({
+          icon: "error",
+          title: "Error",
+          text: "Invalid credentials"
+        });
+      }
     });
   }
   return (
